Extract generic helpers into utils modules
